+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Add selection to Extract Function code action #2118
 - Fix replace-refer-all mangling unrelated symbols #2150
 - Fix unused-public-var :exclude-when-contains-meta not being considered for .lsp/config.edn.
+- Fix unused-public-var not working with :config-in-ns in kondo config.
 
 ## 2025.11.28-12.47.43